make alt-tab able to give priority to windows on the current viewport
(bzr r1629.1.1)

+++ plugins/unityshell/src/LauncherIcon.h
@@ -126,7 +126,7 @@ public:
     return false;
   };
 
-  virtual unsigned int SwitcherPriority()
+  virtual unsigned long long SwitcherPriority()
   {
     return 0;
   }
